+++ src/main/java/com/quaxt/mcc/semantic/Array.java
@@ -6,6 +6,6 @@ public record Array(Type element, Constant arraySize) implements Type {
     @Override
     public boolean looseEquals(Type other) {
         return other instanceof Array(Type otherElement, Constant otherArraySize)
-                && otherElement.looseEquals(otherElement) && arraySize.toInt()==otherArraySize.toInt();
+                && otherElement.looseEquals(otherElement) && arraySize.toLong()==otherArraySize.toLong();
     }
 }
